Move LookAheadStream to .compiler package so that it can be shared by tokenizer and parser
