+++ src/main/java/com/computer/nand2tetris/compiler/tokenizer/IdentifierOrKeywordTokenExtractor.java
@@ -2,6 +2,7 @@ package com.computer.nand2tetris.compiler.tokenizer;
 
 import com.computer.nand2tetris.compiler.JackToken;
 import com.computer.nand2tetris.compiler.JackToken.TokenType;
+import com.computer.nand2tetris.compiler.LookAheadStream;
 import com.google.common.collect.ImmutableSet;
 
 class IdentifierOrKeywordTokenExtractor implements JackTokenExtractor {
