+++ src/main/java/com/computer/nand2tetris/compiler/tokenizer/JackTokenExtractor.java
@@ -1,6 +1,7 @@
 package com.computer.nand2tetris.compiler.tokenizer;
 
 import com.computer.nand2tetris.compiler.JackToken;
+import com.computer.nand2tetris.compiler.LookAheadStream;
 
 interface JackTokenExtractor {
