+++ src/main/java/com/computer/nand2tetris/compiler/tokenizer/StringConstantTokenExtractor.java
@@ -1,6 +1,7 @@
 package com.computer.nand2tetris.compiler.tokenizer;
 
 import com.computer.nand2tetris.compiler.JackToken;
+import com.computer.nand2tetris.compiler.LookAheadStream;
 import com.google.common.base.Preconditions;
 
 class StringConstantTokenExtractor implements JackTokenExtractor {
