+++ src/main/java/com/computer/nand2tetris/compiler/tokenizer/WhitespaceTokenExtractor.java
@@ -2,6 +2,7 @@ package com.computer.nand2tetris.compiler.tokenizer;
 
 import com.computer.nand2tetris.compiler.JackToken;
 import com.computer.nand2tetris.compiler.JackToken.TokenType;
+import com.computer.nand2tetris.compiler.LookAheadStream;
 
 public class WhitespaceTokenExtractor implements JackTokenExtractor {
